ml_requirements.txt
# PyTorch (pinned for GFPGAN/basicsr compatibility)
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# Diffusion and transformers
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
safetensors>=0.4.0
huggingface-hub==0.21.4

# Audio processing
openai-whisper>=20240314
librosa>=0.10.0
soundfile>=0.12.0

# NLP
spacy>=3.7.0

# Image processing
numpy<2
Pillow>=10.0.0
opencv-python>=4.9.0
gfpgan>=1.3.0
facexlib>=0.3.0
basicsr>=1.4.0
realesrgan>=0.3.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# AWS
boto3>=1.30.0

# Environment
python-dotenv>=1.0.0
omegaconf>=2.3.0
